+++ CHANGELOG.md
@@ -9,3 +9,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Build scheme for JourneySquad target.
 - `BillValidationError` type.
+- `Bill` model into the model layer - [#32](https://github.com/ios-course/link-team-project/pull/32)